chore: define proper interface for wasm internals of `Barretenberg`
