+++ acvm_backend_barretenberg/src/barretenberg/mod.rs
@@ -1,0 +1,11 @@
+cfg_if::cfg_if! {
+    if #[cfg(feature = "native")] {
+        mod native;
+
+        pub(crate) use native::Barretenberg;
+    } else {
+        mod wasm;
+
+        pub(crate) use wasm::Barretenberg;
+    }
+}
