+++ acvm_backend_barretenberg/src/barretenberg/native.rs
@@ -1,0 +1,8 @@
+#[derive(Default)]
+pub(crate) struct Barretenberg;
+
+impl Barretenberg {
+    pub(crate) fn new() -> Barretenberg {
+        Barretenberg::default()
+    }
+}
